Refactor Enum codes TileType, BossType, ...
