+++ src/main/java/com/Boss.java
@@ -3,23 +3,22 @@ import lombok.*;
 
 import static com.Validator.isValidForProperty;
 
-enum BossType {
-    AZATHOTH("아자투스"),
-    CHUTHULU("크툴루"),
-    NYARLATHOTEP("니알라토텝");
-
-    final private String name;
-    BossType(String name){
-        this.name = name;
-    }
-    public String getName(){
-        return name;
-    }
-}
-
 @Data
 @Builder
 public class Boss {
+    public enum BossType {
+        AZATHOTH("아자투스"),
+        CHUTHULU("크툴루"),
+        NYARLATHOTEP("니알라토텝");
+
+        final private String name;
+        BossType(String name){
+            this.name = name;
+        }
+        public String getName(){
+            return name;
+        }
+    };
     private BossType name;
     private int health;
     private int power;
