+++ src/test/java/com/MapTest.java
@@ -1,0 +1,13 @@
+package com;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class MapTest {
+
+    @BeforeEach
+    void setUp() {
+    }
+}
